[RUNE-19] Story: ANSI-aware wrap & slice && resolved all SwiftLint violations
