+++ .swiftlint.yml
@@ -50,7 +50,6 @@ opt_in_rules:
   - modifier_order
   - nimble_operator
   - nslocalizedstring_key
-  - number_separator
   - object_literal
   - operator_usage_whitespace
   - overridden_super_call
@@ -144,6 +143,8 @@ type_name:
   max_length:
     warning: 40
     error: 50
+
+
 
 # Custom rules (disabled for initial development)
 # These can be enabled later as the codebase matures
