+++ Sources/RuneComponents/RuneComponents.swift
@@ -1,3 +1,7 @@
+// Re-export main types for convenience
+// Note: Swift doesn't support @_exported import for individual types
+// Types are automatically available when importing the module
+
 /// RuneComponents module - UI components for terminal interfaces
 ///
 /// This module provides a collection of reusable UI components for building
@@ -9,7 +13,4 @@
 /// - Box containers with border styles
 /// - Layout-aware component rendering
 /// - Integration with layout and renderer modules
-
-// Re-export main types for convenience
-// Note: Swift doesn't support @_exported import for individual types
-// Types are automatically available when importing the module
+public enum RuneComponents {}
