+++ Sources/RuneLayout/RuneLayout.swift
@@ -1,16 +1,19 @@
-/// RuneLayout module - Flexbox layout engine for terminal UIs
-///
-/// This module provides a flexbox-inspired layout system optimized for
-/// terminal user interfaces. It handles the arrangement of components
-/// within containers using flexible sizing and alignment.
-///
-/// Key features:
-/// - Flexbox-style layout calculation
-/// - Support for row and column directions
-/// - Justify and align content options
-/// - Constraint-based sizing
-/// - Integration with Unicode width calculations
+// RuneLayout module - Flexbox layout engine for terminal UIs
+//
+// This module provides a flexbox-inspired layout system optimized for
+// terminal user interfaces. It handles the arrangement of components
+// within containers using flexible sizing and alignment.
+//
+// Key features:
+// - Flexbox-style layout calculation
+// - Support for row and column directions
+// - Justify and align content options
+// - Constraint-based sizing
+// - Integration with Unicode width calculations
 
 // Re-export main types for convenience
 // Note: Swift doesn't support @_exported import for individual types
 // Types are automatically available when importing the module
+
+/// RuneLayout module namespace
+public enum RuneLayout {}
